Update factory
Changed the factory class to use a single method for object creation of
multiple types. Removed the private method for calling doTheMath.
Changed to code to use only doTheMath for the calculations. Changed the
subclasses doTheMath to override the inherited variant and do a unique
calculation if their member is set to true / yes.