Corrected types
Changed the operandA, operandB, and answer from NSInteger to float
types. This is for the division math. Cast the other operations as INT
when they are returned. Created a conditional check for the float to
check for a remainder. If no remainder the division answer is cast as a
INT like the others when stringed for return, otherwise it remains a
float type.